refactor: :recycle: modify repo structure
